+++ test/libsolidity/syntaxTests/nameAndTypeResolution/247_multi_variable_declaration_wildcards_fail_6.sol
@@ -1,7 +1,0 @@
-contract C {
-    function two() public returns (uint, uint);
-    function f() public { (uint a, uint b, uint c) = two(); }
-}
-// ----
-// Warning: (87-119): Different number of components on the left hand side (3) than on the right hand side (2).
-// TypeError: (87-119): Not enough components (2) in value to assign all variables (3).