+++ test/libsolidity/syntaxTests/types/unnamed_tuple_decl.sol
@@ -16,4 +16,3 @@ contract C {
 // SyntaxError: (249-261): The use of the "var" keyword is disallowed. The declaration part of the statement can be removed, since it is empty.
 // SyntaxError: (271-283): The use of the "var" keyword is disallowed. The declaration part of the statement can be removed, since it is empty.
 // SyntaxError: (293-306): The use of the "var" keyword is disallowed. The declaration part of the statement can be removed, since it is empty.
-// TypeError: (271-283): Too many components (1) in value for variable assignment (0) needed